+++ SERVE.sh
@@ -1,1 +1,4 @@
-python -m SimpleHTTPServer 8000
+# Install Python 3 before using:
+# cmd: brew install python3 
+
+python3 -m http.server 8000